os: Change os_nanosleep parameter to int64_t

+++ src/xrt/auxiliary/os/os_time.h
@@ -69,7 +69,7 @@ extern "C" {
  * @ingroup aux_os_time
  */
 static inline void
-os_nanosleep(int32_t nsec)
+os_nanosleep(int64_t nsec)
 {
 #if defined(XRT_OS_LINUX)
 	struct timespec spec;
